created component for already did shared form
